Add test cases for List command parsers
